refactor(ListItemByDoc): Move default exp attr creation from children
